black/requirements.txt
paho-mqtt>=1.6.1
smbus2>=0.4.2
RPi.GPIO; sys_platform == 'linux'
netifaces>=0.11.0
